Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -7,7 +7,7 @@ use App\Question;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\DB;
-
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -66,7 +66,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return  view('questions.edit', compact('question'));
+          if (Gate::denies('update-question', $question)){
+              abort(403, 'Access Denied');
+          }
+            return  view('questions.edit', compact('question'));
     }
 
     /**
@@ -78,6 +81,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)){
+            abort(403, 'Access Denied');
+        }
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', 'Your question has been updated.');
     }
@@ -91,6 +97,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('update-question', $question)){
+            abort(403, 'Access Denied');
+        }
         $question->delete();
         return redirect('/questions')->with('success', 'Your question has been deleted.');
     }
